マイページUIの実装_修正
