+++ src/app/api/contact/route.ts
@@ -1,7 +1,7 @@
-import { NextRequest, NextResponse } from 'next/server';
-import { PrismaClient } from '@prisma/client';
-import { Slack } from '@/app/_utils/Slack';
-import { SendMail } from '@/app/_hooks/SendMail';
+import { NextRequest, NextResponse } from "next/server";
+import { PrismaClient } from "@prisma/client";
+import { Slack } from "@/app/_utils/Slack";
+import { SendMail } from "@/app/_hooks/SendMail";
 
 const prisma = new PrismaClient();
 
@@ -16,25 +16,27 @@ export const POST = async (request: NextRequest) => {
         name,
         email,
         content,
-      }
+      },
     });
 
     // スラックに通知を送信
-    await Slack(name, email, content); 
+    await Slack(name, email, content);
 
     // 問い合わせ完了メール送信
-    await SendMail(name, email, content); 
+    await SendMail(name, email, content);
 
     // 成功
     return NextResponse.json({
-      status: 'OK',
-      message: 'お問い合わせ完了',
+      status: "OK",
+      message: "お問い合わせ完了",
       id: contact.id,
     });
-
   } catch (error) {
     if (error instanceof Error) {
-      return NextResponse.json({ status: 'error', message: error.message }, { status: 400 });
+      return NextResponse.json(
+        { status: "error", message: error.message },
+        { status: 400 }
+      );
     }
-  };
+  }
 };
