+++ src/app/mypage/_hooks/useRouteGuard.ts
@@ -1,6 +1,6 @@
-import { useSupabaseSession } from '@/app/_hooks/useSupabaseSession';
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
+import { useRouter } from "next/navigation";
+import { useEffect } from "react";
 
 export const useRouteGuard = () => {
   const router = useRouter();
@@ -11,8 +11,8 @@ export const useRouteGuard = () => {
 
     const fetcher = async () => {
       if (session === null) {
-        router.replace('/login')
-      };
+        router.replace("/login");
+      }
     };
 
     fetcher();
